refactor: unify component directories and rename auth pages
