+++ src/app/sign_up/page.tsx
@@ -5,9 +5,9 @@ import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { supabase } from '@/lib/supabaseClient';
 import { useUserStore } from '@/store/userStore';
-import { Button } from '@/component/ui/button';
-import { Input } from '@/component/ui/input';
-import { Label } from '@/component/ui/label';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
 import { toast } from 'sonner';
 
 export default function SignUpPage() {
@@ -88,7 +88,7 @@ export default function SignUpPage() {
           </Button>
           <p className="text-sm text-center">
             Already have an account?{' '}
-            <Link href="/signin" className="underline">
+            <Link href="/sign_in" className="underline">
               Sign in
             </Link>
           </p>
